- components folder renamed to presentation
