+++ src/lib.rs
@@ -22,7 +22,7 @@
 
 mod application;
 mod common;
-mod components;
+mod presentation;
 mod domain;
 mod infrastructure;
