+++ src/presentation/molecules/section.rs
@@ -23,8 +23,8 @@
 use leptos::prelude::*;
 
 use crate::common::size::*;
-use crate::components::atoms::layout::{spacing, Align, Stack};
-use crate::components::atoms::typography::{H1, H2, H3, H4};
+use crate::presentation::atoms::layout::{spacing, Align, Stack};
+use crate::presentation::atoms::typography::{H1, H2, H3, H4};
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 #[allow(dead_code)]
