+++ src/presentation/organisms/job_config_form.rs
@@ -26,13 +26,13 @@ use std::str::FromStr;
 use uuid::Uuid;
 
 use crate::common::size::*;
-use crate::components::atoms::alert::*;
-use crate::components::atoms::button::*;
-use crate::components::atoms::layout::*;
-use crate::components::atoms::typography::*;
-use crate::components::molecules::form_field::*;
-use crate::components::molecules::schema_form::{SchemaForm, SchemaFormState};
-use crate::components::molecules::section::*;
+use crate::presentation::atoms::alert::*;
+use crate::presentation::atoms::button::*;
+use crate::presentation::atoms::layout::*;
+use crate::presentation::atoms::typography::*;
+use crate::presentation::molecules::form_field::*;
+use crate::presentation::molecules::schema_form::{SchemaForm, SchemaFormState};
+use crate::presentation::molecules::section::*;
 use crate::domain::application_info::AppInfo;
 use crate::domain::job_context::*;
 use crate::domain::runner_info::RunnerInfo;
